Split util files
